RDCC-6040:Changes made to incorporate review comments
